requirements.txt
# Core dependencies
requests==2.31.0
pandas==2.2.2
numpy==1.26.4
pyyaml==6.0.1

# Graph ML and evaluation
pykeen==1.10.1
torch>=2.0.0
scikit-learn>=1.3.0

# Optional for visualisation or graph processing
networkx==3.3
matplotlib==3.8.4

# Jupyter (for notebooks)
jupyterlab==4.1.8

# Development and testing
pytest>=7.0.0
